[Refactor]dto위치 변경
