+++ src/main/java/com/codeit/weatherwear/domain/clothes/dto/request/ClothesAttributeDto.java
@@ -1,4 +1,4 @@
-package com.codeit.weatherwear.domain.clothes.dto.response;
+package com.codeit.weatherwear.domain.clothes.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
