+++ src/main/java/com/codeit/weatherwear/domain/clothes/dto/request/ClothesCreateRequest.java
@@ -1,6 +1,5 @@
 package com.codeit.weatherwear.domain.clothes.dto.request;
 
-import com.codeit.weatherwear.domain.clothes.dto.response.ClothesAttributeDto;
 import com.codeit.weatherwear.domain.clothes.entity.ClothType;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
