+++ src/main/java/com/codeit/weatherwear/domain/clothes/service/ClothServiceImpl.java
@@ -1,6 +1,6 @@
 package com.codeit.weatherwear.domain.clothes.service;
 
-import com.codeit.weatherwear.domain.clothes.dto.response.ClothesAttributeDto;
+import com.codeit.weatherwear.domain.clothes.dto.request.ClothesAttributeDto;
 import com.codeit.weatherwear.domain.clothes.dto.request.ClothesCreateRequest;
 import com.codeit.weatherwear.domain.clothes.dto.request.ClothesUpdateRequest;
 import com.codeit.weatherwear.domain.clothes.dto.response.ClothesDto;
